SPY-157: Add Auth logging.
Motivation
----------
Debugging how long authentication steps take is currently not
possible. Proper logging needs to be added in order to debug
authentication timeframes.

Modifications
-------------
Added logging infrastructure for the individual steps (list sasl
mechs, and challenge/response cycles), as well as a total amount
on a per-node basis.

If an individual step takes over 250ms it is WARNed instead of
DEBUGed, and if the total auth step takes over 1 second it it also
WARNed. Note that while these timings are fixed values, in general
the operation taking this long points to something wrong since it
is expected to take in the 1ms range.

Result
------
Auth times can now be observed both in development and production,
better pointing to possible issues.

Change-Id: I3deab53332251263c5a6dcf3e2bfd75538dfb24c
Reviewed-on: http://review.couchbase.org/34849
Reviewed-by: Matt Ingenthron <matt@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>