SPY-160: Redistribute ops when waiting for authentication to complete.
Motivation
----------
When an operation gets added and the node needs to (re)connect, it waits
for a given amount of time (authWaitTime). If the auth does not complete
in the period of time, the operation is cancelled. This is suboptimal,
especially if the operation is not timed out yet, it could wait longer
until it gets timed out/cancelled.

Modifications
-------------
This changeset makes sure that an operation gets redistributed (assuming
FailureMode is NOT cancel) if the authLatch is not counted down in the
wait period.

Redistributing will only happen (further down the code stack) if the
op is 1) not cancelled and 2) not timed out already.

Result
------
Still valid operations will be transparently rescheduled until they
either go through validly or will eventually time out or get
cancelled.

Change-Id: I6f849314e11435a6ebc69a4bfc1eab705df850e8
Reviewed-on: http://review.couchbase.org/34852
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>