SPY-162: Also respect nagle setting on reconnect.
Motivation
----------
If a server connection is reset (because it is lost or a reconnection
is triggered), the nagle setting is not re-set properly.

Modifications
-------------
During the reconnect phase, properly re-set the nagle setting so it is
respected on the new connection.

Result
------
The custom nagle setting is respected and used even across new connections.

Change-Id: Ic0b513b8538a39710aaf8fd0f0f699a5c561bc09
Reviewed-on: http://review.couchbase.org/35553
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Matt Ingenthron <matt@couchbase.com>