Enhanced auth logging on sasl failure.
Motivation:
-----------
In case of long-taking auth roundtrips, it could be the case - for one 
reason or another - that a empty sasl list gets returned or the auth 
thread gets interrupted, leading to the same effect. This needs to 
be better debugable from the logs.

Modifications:
--------------
Better logging has been added for auth failure scenarios, and also the 
sasl mechs response has been refactored a bit so that the timing is 
also reported in case of an error.

Result:
-------
Auth errors are now easier to debug from the logs.

Change-Id: I9f0fff8b2c3a6b1492fb0fd82746ff78f7b006c5
Reviewed-on: http://review.couchbase.org/35672
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Matt Ingenthron <matt@couchbase.com>