SPY-163: Count down bulk get latch even when empty key list is provided.
Motivation
----------
When a empty iterator (or empty key list) is passed in to a get bulk operation,
a result is never returned and the thread is blocked because it waits on a
latch that will never be counted down (because no response will ever arrive).

Modifications
-------------
If no chunks with keys are sent out to the servers, the latch is initialized
to 0 right away so that the code doesn't need to wait at all.

Result
------
The code now returns properly with an empty map instead of blocking the thread
forever.

Change-Id: I0711c399d9f15010bc808e0f651ce6ad605c06f0
Reviewed-on: http://review.couchbase.org/36208
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Matt Ingenthron <matt@couchbase.com>