SPY-169: Also clone observe calls on redistribute.
Motivation
----------
Observe calls also need to be cloned when redistributed, for example if a socket
gets closed or the op stil waits for an auth latch.

Modifications
-------------
The observe operation now gets cloned similar to get, gets and so on.

Result
------
No assert error is raised if an observe needs to be cloned.

Change-Id: I6269a1d7fb756855f94caa2fe0495133fe1a83b0
Reviewed-on: http://review.couchbase.org/36698
Reviewed-by: Matt Ingenthron <matt@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>