SPY-170: Fix concurrency issue in StringUtils.isJSONObject()
Motivation
----------
A concurrency issue was reported in the StringUtils class, pointing towards
the pattern matching of the JSON string.

Modifications
-------------
It turned out that the matcher is not thread safe and there is a race condition
between clearing it and then matching on it. By falling back to the thread-safe
builder directly the race condition goes away.

Result
------
No concurrency issue anymore in the helper method.

Change-Id: I09729af78bd241ff8dcb0869992d3ff474fb6ec1
Reviewed-on: http://review.couchbase.org/37603
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>