SPY-134: Make sure lastReadTimestamp is concurrency safe.
Motivation
----------
When the last read timestamp is read from each node, there has been some
suspicion that the (sometimes very long) reported times are flawed.

Tracking it down further, it was discovered that the timestamp was potentially
accessed by two or more threads, but not volatile.

Modifications
-------------
This changeset makes the timstamp volatile and also uses nanoTime instead
of currentTimeMillis for better accuracy.

Result
------
Since the timestamp is now thread safe, the actual value will now be shown
to the user.

Change-Id: I39e47408319f8adf3b2875a4a730232ecadeafe2
Reviewed-on: http://review.couchbase.org/37646
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Matt Ingenthron <matt@couchbase.com>