Move superfluous log message from info to debug.
Motivation
----------
In every log when connecting, a message like this shows up:

INFO [Memcached IO over {MemcachedConnection to ...}] (?:?) - 
Connection state changed for sun.nio.ch.SelectionKeyImpl@...

This message is superfluous since it does not provide any addition 
information to the user, especially not at INFO level.

Modifications
-------------
The message has been moved to DEBUG level so that users are not 
wondering what it means and also to streamline the log a bit.

Result
------
Easier log output to read and analyze for the user.

Change-Id: I13ef691dd435f397dc9d5f08ff40a28202d3ddb7
Reviewed-on: http://review.couchbase.org/37647
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Matt Ingenthron <matt@couchbase.com>