Change MemcachedNodeROImpl and BaseGetOpImpl classes to public.
Motivation
----------
Implementing a custom NodeLocator requires to override the getReadonlyCopy
method which in turn requires to use some MemcachedNode implementation.
MemcachedNodeROImpl should be reusable outside of its package for that
purpose. In addition, memcached being an open source project, its
implementation can be changed or extended and commands such as GET may
be used differently in these custom versions. The abstract BaseGetOpImpl
should be flexible enough to be adapted to different but similar GET
commands.

Modifications
-------------
The MemcachedNodeROImpl class accessor changes from package-private to public.
The BaseGetOpImpl class accessor changes from package-private to public.
The initialization method and constructors of BaseGetOpImpl changed to
abstract the string appended to the <GET key> command.

Result
------
Writing a custom NodeLocator outside of the net.spy.memcached package
and supporting variants of GET commands is now easy.

Change-Id: I49b3efc741ab4fe5780bf74ad5f99b839c9ceb7d
Reviewed-on: http://review.couchbase.org/44565
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>