SPY-181: GetAndLock needs to be cloneable like any other keyed op.
Motivation
----------
In case a getAndLock operation needs to be rescheduled, it needs to
be cloneable (like any other keyed operation).

Modifications
-------------
Apply the same clone logic as with any other keyed operation. Also
added getter methods to the operations so that the expiration time
can be extracted on cloning.

Result
------
Correct behavior when a getAndLock op needs to be cloned.

Change-Id: I7f4fdad62af71a42cc203b2a4b7e72f2002a286d
Reviewed-on: http://review.couchbase.org/45129
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>