SPY-182: Avoid WARN logging when handling node not set.
Motivation
----------
The main purpose of setting a continous timeout on a node is to
properly reconnect when the threshold is reached. Now if auth
has not happened yet, it could be that ops do not have a handling
node set yet.

In those cases, it doesn't make any sense to increase the counter,
so it should be ignored. This also has the side effect of not logging
warnings to the user which have no impact for him.

Modifications
-------------
Ignore the operation when its handling node is not set.

Result
------
No unnecessary warnings printed to the user.

Change-Id: I3643d5e0bfc7474889279d0f8362f287624b89a3
Reviewed-on: http://review.couchbase.org/45131
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>