SPY-183: Allow touch operations to be cloned.
Motivation
----------
In case a touch  operation needs to be rescheduled, it needs to
be cloneable (like any other keyed operation).

Modifications
-------------
Apply the same clone logic as with any other keyed operation. Also
added getter methods to the operations so that the expiration time
can be extracted on cloning.

Result
------
Correct behavior when a touch op needs to be cloned.

Change-Id: Iad343b4dbdcd5dfd4d9ec53bf5335dcf4775f9c9
Reviewed-on: http://review.couchbase.org/48076
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>