SPY-191: Don't force reconnect on E2BIG with binary protocol.
Motivation
----------
In the current codebase, a socket is forcefully reset when the server
returns an E2BIG response, that is when the document is larger than the
possible value size.

Modifications
-------------
While this needs to be done on the ASCII protocol, doing so is not needed
with binary protocol and can be considered a bug/leftover.

The code is modified so that E2BIG just translates into a non-success response
like any other error and the proper error code identifies the cause.

The test cases have been modified for binary to reflect the change.

Result
------
Do not force reconnect / treat the issue as a server error on the binary
protocol - the ASCII protocol is unaffected.

Change-Id: Idf392d146d30b2e96dc198a93a3cc6598dae3fc6
Reviewed-on: http://review.couchbase.org/61927
Reviewed-by: Simon Baslé <simon@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>