SPY-194: Allow to optionally bound retry queue.
Motivation
----------
If the retry queue might grow very large this can lead to
uncontrollable side effects (like heap growth), since this
queue is not accessible from the outside of spymemcached.

Modifications
-------------
This change adds the system property "net.spy.retryQueueSize" which
allows to optionally bound the retry queue to the given size.

Note that to maximize backwards compatibility and to not introduce
any regressions the actual queue has not been changed but rather
the code which inserts into this queue is now checking for the
limit.

If the queue is full the operation is cancelled instead of sending
it into the retry queue.

Result
------
It is now possible to bound the retry queue.

Change-Id: I86452ef8dcddcf2c4acfc065fd9db94510665ac3
Reviewed-on: http://review.couchbase.org/69938
Tested-by: Michael Nitschinger <michael@nitschinger.at>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>