SPY-193: Allow to override wakeupDelay
Motivation
----------
As part of SPY-172, a new wakeupDelay setting has been introduced,
but as part of a mistake it has never been used to actually set
the wakeup time, which is still hardcoded to 1000.

Modifications
-------------
This simple change modifies the code so it accepts the wakeupDelay
setting, still having the default at 1000 but making it possible
to override it through the system property.

Result
------
The wakeupDelay can now be configured via a system property.

Change-Id: If1273cb560d49ea2c9c42b7c7b70c84fd9aeebdc
Reviewed-on: http://review.couchbase.org/75116
Tested-by: Michael Nitschinger <michael@nitschinger.at>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>